split nat64 to its own package
