+++ pkg/network/utils.go
@@ -1,0 +1,73 @@
+package network
+
+import (
+	"fmt"
+	"net"
+
+	"github.com/vishvananda/netlink"
+	"k8s.io/klog/v2"
+)
+
+// GetMTU returns the MTU used for the IP family
+func GetMTU(ipFamily int) (int, error) {
+	iface, err := GetDefaultGwInterface(ipFamily)
+	if err != nil {
+		return 0, err
+	}
+	mtu, err := getInterfaceMTU(iface)
+	if err != nil {
+		return 0, err
+	}
+	return mtu, nil
+}
+
+// getInterfaceMTU finds the mtu for the interface
+func getInterfaceMTU(iface string) (int, error) {
+	interfaces, err := net.Interfaces()
+	if err != nil {
+		return 0, err
+	}
+	for _, inter := range interfaces {
+		if inter.Name == iface {
+			return inter.MTU, nil
+		}
+	}
+	return 0, fmt.Errorf("no %s device found", iface)
+}
+
+func GetDefaultGwInterface(ipFamily int) (string, error) {
+	routes, err := netlink.RouteList(nil, ipFamily)
+	if err != nil {
+		return "", err
+	}
+
+	for _, r := range routes {
+		// no multipath
+		if len(r.MultiPath) == 0 {
+			if r.Gw == nil {
+				continue
+			}
+			intfLink, err := netlink.LinkByIndex(r.LinkIndex)
+			if err != nil {
+				klog.Infof("Failed to get interface link for route %v : %v", r, err)
+				continue
+			}
+			return intfLink.Attrs().Name, nil
+		}
+
+		// multipath, use the first valid entry
+		// xref: https://github.com/vishvananda/netlink/blob/6ffafa9fc19b848776f4fd608c4ad09509aaacb4/route.go#L137-L145
+		for _, nh := range r.MultiPath {
+			if nh.Gw == nil {
+				continue
+			}
+			intfLink, err := netlink.LinkByIndex(r.LinkIndex)
+			if err != nil {
+				klog.Infof("Failed to get interface link for route %v : %v", r, err)
+				continue
+			}
+			return intfLink.Attrs().Name, nil
+		}
+	}
+	return "", fmt.Errorf("not routes found")
+}
